+++ tests/extension/src/type_checking_aliased/foo.py
@@ -1,0 +1,4 @@
+class Foo:
+    """Foo test class."""
+
+    attr: str = "test"